// src/api/generated/model/reasonEnum.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * SurgiScribe API Documentation
 * API documentation with JWT Authentication
 * OpenAPI spec version: v1
 */

/**
 * * `surgery` - Surgery
 * * `restock` - Restock
 * * `return` - Return
 * * `other` - Other
 */
export type ReasonEnum = typeof ReasonEnum[keyof typeof ReasonEnum];


export const ReasonEnum = {
  surgery: 'surgery',
  restock: 'restock',
  return: 'return',
  other: 'other',
} as const;