// src/api/generated/model/transferTargetTypeEnum.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * SurgiScribe API Documentation
 * API documentation with JWT Authentication
 * OpenAPI spec version: v1
 */

/**
 * * `representative` - Representative
 * * `facility` - Facility
 */
export type TransferTargetTypeEnum = typeof TransferTargetTypeEnum[keyof typeof TransferTargetTypeEnum];


export const TransferTargetTypeEnum = {
  representative: 'representative',
  facility: 'facility',
} as const;