// src/api/generated/model/transportMethodEnum.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * SurgiScribe API Documentation
 * API documentation with JWT Authentication
 * OpenAPI spec version: v1
 */

/**
 * * `fedex` - FedEx
 * * `ups` - UPS
 * * `rep` - Rep
 * * `other` - Other
 */
export type TransportMethodEnum = typeof TransportMethodEnum[keyof typeof TransportMethodEnum];


export const TransportMethodEnum = {
  fedex: 'fedex',
  ups: 'ups',
  rep: 'rep',
  other: 'other',
} as const;